Corrected connectify shim
The shim drops the additional arguments which can be passed to connect.  The official connect method signature includes output and input numbers as well as the node to connect to.  Using in it's current state causes all audio to come out of the left channel when using certain effects.

This update also addresses the issue of creating multiple Tuna instances in a session.  The previous connectify call would create a chain of shims, adding one each time new Tuna(...) is called.